Crud wishlist

+++ models/GiftsModel.php
@@ -30,7 +30,8 @@ class GiftsModel extends Database {
     }    
     
     public function addGift($gift_title, $gift_description, $gift_link, $gift_image, $wishlist_id) {
-        $sql = "INSERT INTO gift(gift_title, gift_description, gift_link, gift_image, wishlist_id) VALUES (:gift_title, :gift_description, :gift_link, :gift_image, :wishlist_id)";
+        $sql = "INSERT INTO gift(gift_title, gift_description, gift_link, gift_image, wishlist_id) 
+        VALUES (:gift_title, :gift_description, :gift_link, :gift_image, :wishlist_id)";
         $stmt = $this->connect()->prepare($sql);
         $stmt->bindParam(":gift_title", $gift_title);
         $stmt->bindParam(":gift_description", $gift_description);
@@ -38,24 +39,24 @@ class GiftsModel extends Database {
         $stmt->bindParam(":gift_image", $gift_image);
         $stmt->bindParam(":wishlist_id", $wishlist_id);
         $stmt->execute();
-        $addGift = $stmt->fetchAll(PDO::FETCH_ASSOC);
         $stmt->closeCursor();
-        return $addGift;
+        return true;
     }
 
-    public function getGiftsByWishlist($wishlist_id) {
-        $query = "SELECT
+/*     public function getListDetailById($wishlist_id) {
+        $sql = "SELECT
         gift.*,
         wishlist.wishlist_id
         FROM gift JOIN wishlist ON gift.wishlist_id = wishlist.wishlist_id
         WHERE gift.wishlist_id = :wishlist_id";
-        $stmt = $this->connect()->prepare($query);
-        $stmt->bindParam(":wishlist_id", $wishlist_id);
+        $stmt = $this->connect()->prepare($sql);
+        $stmt->bindParam(":wishlist_id", $wishlist_id, PDO::PARAM_INT);
         $stmt->execute();
         $giftsByWishlist = $stmt->fetchAll(PDO::FETCH_ASSOC);
         $stmt->closeCursor();
         return $giftsByWishlist;
-    }
+
+    } */
 
     public function getGiftById($gift_id) {
         $query = "SELECT * FROM gift WHERE gift_id = :gift_id";
@@ -65,14 +66,31 @@ class GiftsModel extends Database {
         return $stmt->fetchAll(PDO::FETCH_ASSOC);
     }
 
-    public function deleteGift($gift_id) {
-        $query = "DELETE FROM gift WHERE gift_id = :gift_id";
+    public function deleteGiftById($gift_id, $user_id) {
+        $query = "DELETE FROM gift WHERE gift_id = :gift_id AND wishlist_id IN (SELECT wishlist_id FROM wishlist WHERE user_id = :user_id)";
         $stmt = $this->connect()->prepare($query);
-        $stmt->bindParam(":gift_id", $gift_id);
-        return $stmt->execute();
+        $stmt->bindParam(":gift_id", $gift_id, PDO::PARAM_INT);
+        $stmt->bindParam(":user_id", $user_id, PDO::PARAM_INT);
+        $stmt->execute();
+        $stmt->closeCursor();
+        return true;    
     }
-}
-    /* public function getListDetail() {
+    public function getGiftsByWishlist() {
+        $query = "SELECT
+        gift.*,
+        wishlist.wishlist_id
+        FROM gift JOIN wishlist ON gift.wishlist_id = wishlist.wishlist_id
+        WHERE gift.wishlist_id = :wishlist_id";
+        $stmt = $this->connect()->prepare($query);
+        $stmt->bindParam(":wishlist_id", $wishlist_id, PDO::PARAM_INT);
+        $stmt->execute();
+        $giftsByWishlist = $stmt->fetchAll(PDO::FETCH_ASSOC);
+        $stmt->closeCursor();
+        return $giftsByWishlist;
+    }
+
+
+/*      public function getListDetail() {
         $sql = "SELECT * FROM gift WHERE wishlist_id = :wishlist_id";
         $stmt = $this->connect()->prepare($sql);
         $stmt->bindParam(":wishlist_id", $wishlist_id);
@@ -80,18 +98,16 @@ class GiftsModel extends Database {
         $gifts = $stmt->fetchAll(PDO::FETCH_ASSOC);
         $stmt->closeCursor();
         return $gifts;
-    }
+    } */
 
-
-
-    public function getGiftsByWishlist($wishlist_id) {
-        $query = "SELECT * FROM " . $this->table_name . " WHERE wishlist_id = :wishlist_id";
-        $stmt = $this->conn->prepare($query);
-        $stmt->bindParam(":wishlist_id", $wishlist_id);
+/*     public function getGiftsByWishlist($wishlist_id) {
+        $sql = "SELECT * FROM gift WHERE wishlist_id = :wishlist_id";
+        $stmt = $this->connect()->prepare($sql);
+        $stmt->bindParam(":wishlist_id", $wishlist_id, PDO::PARAM_INT);
         $stmt->execute();
-        return $stmt->fetchAll(PDO::FETCH_ASSOC);
-    }*/
-
-
-
+        $datas = $stmt->fetchAll(PDO::FETCH_ASSOC);
+        $stmt->closeCursor();
+        return $datas;
+    } */
+}
 ?>
